Adds anthropic /v1/messages endpoint to openai api_server (#27882)
